submodule model, split model in several files
