+++ fmskill/model/__init__.py
@@ -1,0 +1,5 @@
+from .factory import ModelResult
+from .dfs import DfsModelResult, DfsModelResultItem
+from .pandas import DataFrameModelResult, DataFrameModelResultItem
+
+# from .abstract import ModelResultInterface, MultiItemModelResult
